Small chenges
Purged Player movement script.
Swapped codes in PlayerJump and PlayerJumpOptimised scripts.
Set some defaults.
