+++ Brackeys Game Jam 2020.1/Assets/Scenes/Wesley's Scene.unity
@@ -258,8 +258,8 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: a7e3e8f8520278b68b5d59bf8b8c9565, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  moveSpeed: 4
-  rb: {fileID: 427076969}
+  moveSpeed: 3
+  rb: {fileID: 0}
 --- !u!1 &1334553872
 GameObject:
   m_ObjectHideFlags: 0
