+++ Brackeys Game Jam 2020.1/Assets/Scripts/Player/PlayerOptimisedJump.cs
@@ -2,16 +2,26 @@
 using System.Collections.Generic;
 using UnityEngine;
 
-public class PlayerOptimisedJump : MonoBehaviour
+public class PlayerJump : MonoBehaviour
 {
     [Range(1, 10)]
-    public float jumpVelocity;
+    public float fallMultiplier = 3f;
+    public float lowJumpMultiplier = 2f;
+    
+    Rigidbody2D rb;
+//    
+    void Awake(){
+        rb = GetComponent<Rigidbody2D>();
+    }
     
     void Update()
     {
-        if (Input.GetButtonDown("Jump"))
+        if (rb.velocity.y <0)
         {
-            GetComponent<Rigidbody2D>().velocity = Vector2.up * jumpVelocity;
+            rb.velocity += Vector2.up * Physics2D.gravity.y * (fallMultiplier - 1) * Time.deltaTime;
+        }   else if (rb.velocity.y > 0 && !Input.GetButton("Jump")){
+            rb.velocity += Vector2.up * Physics2D.gravity.y * (lowJumpMultiplier - 1) * Time.deltaTime;
         }
+        
     }
 }
